Rename table to data-table
