+++ frontend/src/components/organization/org-secrets-table.tsx
@@ -13,6 +13,7 @@ import {
   DropdownMenuItem,
   DropdownMenuTrigger,
 } from "@/components/ui/dropdown-menu"
+import { DataTable, DataTableColumnHeader } from "@/components/data-table"
 import { CenteredSpinner } from "@/components/loading/spinner"
 import { AlertNotification } from "@/components/notifications"
 import {
@@ -23,7 +24,6 @@ import {
   UpdateOrgSecretDialog,
   UpdateOrgSecretDialogTrigger,
 } from "@/components/organization/org-secret-update"
-import { DataTable, DataTableColumnHeader } from "@/components/table"
 
 export function OrgSecretsTable() {
   const { orgSecrets, orgSecretsIsLoading, orgSecretsError } = useOrgSecrets()
