+++ frontend/src/components/workspaces/workspace-members-table.tsx
@@ -47,7 +47,7 @@ import {
   DataTable,
   DataTableColumnHeader,
   type DataTableToolbarProps,
-} from "@/components/table"
+} from "@/components/data-table"
 
 export function WorkspaceMembersTable({
   workspace,
